updated code for testing -commit 1

+++ include/thruster_interface.hpp
@@ -3,6 +3,6 @@
 #include <Arduino.h>
 #include <Servo.h> 
 void initializeThrusters();
-void setThrusterThrottle(const int8_t *pwm_values);
+void setThrusterThrottle(const int32_t *pwm_values);
 // void setThrusterThrottles(const int8_t *throttles);
 #endif // THRUSTER_INTERFACE_HPP
